+++ Packages/com.pidan-workshop.unity-zed/Editor/ZedEditorPrefs.cs
@@ -7,7 +7,6 @@ namespace Unity.Zed.Editor
         private const string Prefix = "ZedEditor_";
 
         public const string ExecutablePathKey = Prefix + "ExecutablePath";
-        public const string AdditionalArgsKey = Prefix + "AdditionalArgs";
 
         public static string GetExecutablePath()
         {
@@ -17,16 +16,6 @@ namespace Unity.Zed.Editor
         public static void SetExecutablePath(string path)
         {
             EditorPrefs.SetString(ExecutablePathKey, path ?? "");
-        }
-
-        public static string GetAdditionalArgs()
-        {
-            return EditorPrefs.GetString(AdditionalArgsKey, "");
-        }
-
-        public static void SetAdditionalArgs(string args)
-        {
-            EditorPrefs.SetString(AdditionalArgsKey, args ?? "");
         }
     }
 }